# openapi.yml
...
openapi: 3.1.0
info:
  title: Postmill API (reaatech fork)
  version: 3.4.0
  description: |
    Comprehensive OpenAPI description of the **Postmill** backend (NestJS) as it exists in this
    heavily-modified `reaatech` fork.

    Postmill schedules social-media and chat posts to 28+ channels, with a calendar, analytics, team
    management, a media library, and — in this fork — a pluggable multi-provider **AI layer**,
    **Analytics v2** persisted dashboards, **social comments** sync/moderation, admin-configurable
    **channel OAuth credentials**, and an OAuth2 authorization-code provider.

    ## Surfaces & authentication

    There are three distinct authentication surfaces:

    1. **Dashboard API** (the bulk of routes). Requires a Postmill user session JWT supplied either as
       the `auth` **cookie** or the `auth` **request header**. The active organization is selected
       with the `showorg` cookie/header. Many routes are further gated by RBAC policies
       (`@CheckPolicies`) or by `isSuperAdmin`.
    2. **Public integration API** (`/public/v1/*`). Requires an organization **API key** (or a
       `pos_`-prefixed OAuth access token) in the `Authorization` header. Intended for
       n8n/Zapier/CLI/3rd-party automation.
    3. **Unauthenticated routes** — registration/login, the OAuth2 authorization/token endpoints,
       Stripe & enterprise webhooks, the social-connect OAuth callback, health/monitor, and public
...

...

    ## Rate limiting
...

  contact:
...
  license:
...
servers:
...
security:
  - cookieAuth: []
  - authHeader: []
paths:
...
  /copilot/chat:
    post:
      tags: [Copilot]
      summary: CopilotKit chat runtime (GraphQL-over-HTTP)
      operationId: copilotChat
      description: |
        Proxies the CopilotKit Node HTTP runtime. Body/response follow the CopilotKit GQL protocol.
        Short-circuits with `500` if no AI provider (admin config or `OPENAI_API_KEY`) is available.
      requestBody:
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: CopilotKit runtime response }
        '422': { description: AI provider not configured }
  /copilot/agent:
    post:
      tags: [Copilot]
      summary: Mastra agent runtime (CopilotKit)
      operationId: copilotAgent
      requestBody:
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: Agent runtime response }
        '422': { description: AI provider not configured }
  /copilot/credits:
    get:
      tags: [Copilot]
      summary: Remaining AI credits
      operationId: copilotCredits
      parameters:
        - { name: type, in: query, schema: { type: string, enum: [ai_images, ai_videos], default: ai_images } }
      responses: { '200': { description: Credits } }
  /copilot/list:
    get:
      tags: [Copilot]
      summary: List chat threads
      operationId: copilotList
      parameters: [{ name: perPage, in: query, schema: { type: integer, default: 50 } }]
      responses:
        '200':
          description: Threads
          content:
            application/json:
              schema:
                type: object
                properties:
                  threads:
                    type: array
                    items:
                      type: object
                      properties: { id: { type: string }, title: { type: string } }
  /copilot/{thread}/list:
    get:
      tags: [Copilot]
      summary: List messages in a thread
      operationId: copilotThreadList
      parameters: [{ name: thread, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Messages } }

  # ===================================================================== ENTERPRISE (signed webhooks)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Session JWT in the `auth` cookie (dashboard API).
    authHeader:
      type: apiKey
      in: header
      name: auth
      description: Session JWT in the `auth` header (alternative to the cookie).
    apiKey:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Organization **API key** or a `pos_`-prefixed OAuth access token, sent **raw** (no `Bearer`
        prefix) in the `Authorization` header. Used by `/public/v1/*`.
...